record mem blocks in Spike Event

+++ cosim/emulator/src/util.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 /// @return: binary[a, b]
-inline uint32_t clip(uint32_t binary, int a, int b) { return (binary >> a) & ((1 << (b - a + 1)) - 1); }
+inline uint64_t clip(uint64_t binary, int a, int b) { return (binary >> a) & ((1 << (b - a + 1)) - 1); }
 
 inline bool is_vector_instr(uint64_t f) {
   uint32_t opcode = clip(f, 0, 6);
